services/files_management/requirements.txt
fastapi
uvicorn
python-dotenv
minio
aiofiles
biopython